refactor(auth): split auth.py into a package by responsibility
350줄 단일 모듈을 책임별 패키지(app/core/auth/)로 분리:
- model.py        CurrentUser (인증 주체 모델)
- client.py       OIDCClient / oidc_client (discovery·JWKS·토큰 검증)
- dependencies.py get_current_user / get_optional_current_user / get_current_user_synced
- middleware.py   AuthMiddleware
- __init__.py     공개 심볼 재export → `from app.core.auth import X` 하위호환 유지

외부 import는 전부 `from app.core.auth import X`라 무변경. 단, tests/core/test_auth.py가
`patch("app.core.auth.settings")`로 패치하던 것을 분리된 위치(client/dependencies)로 갱신
("patch where it's looked up"). 순환 import 없음, get_current_user_synced의 지연 import 유지.

869 tests pass. import app.main OK.
